+++ CHANGELOG.md
@@ -1,5 +1,84 @@
 # EDGEGRID GOLANG RELEASE NOTES
 
+## X.X.X (X X, X)
+
+#### FEATURES/ENHANCEMENTS:
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+#### BUG FIXES:
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 ## 8.0.0 (March 19, 2024)
 
 #### BREAKING CHANGES: